*Flyttet Unleash kode for skriv til oss forside til forside-komponenten *Refaktorering
