+++ src/language/nb.ts
@@ -26,11 +26,12 @@ export default {
     "Tjenesten er dessverre ikke tilgjengelig for øyeblikket.",
   "skrivtiloss.tittel": "Skriv til oss",
   "skrivtiloss.description":
-    "Her kan du sende en beskjed om endringer som har betydning for saken og oppfølgingen din fra NAV. Normal svartid er 2 arbeidsdager.",
+    "Her kan du sende en beskjed om endringer som har betydning for saken og oppfølgingen din fra NAV. Normal svartid" +
+    " er 2 arbeidsdager.",
   "skrivtiloss.svartid":
     "Du kan sende oss spørsmål eller informasjon om saken din. Du får normalt svar innen to arbeidsdager.",
   "skrivtiloss.svartid.lang":
-    " På grunn av stor pågang kan svartid for tiden være noe lengre. Vi beklager dette.",
+    " På grunn av stor pågang kan svartiden akkurat nå være noe lengre. Vi beklager dette.",
   "skrivtiloss.ingress":
     "Du må skrive til oss på norsk eller engelsk. Du kan ikke kontakte oss på vegne av andre. Det gjelder også" +
     " fullmektig og verge.",
@@ -45,17 +46,17 @@ export default {
     "Sykemeldt, sykepenger, arbeidsavklaringspenger, pleiepenger, yrkesskade/ménerstatning, hjelp og veiledning for å" +
     " komme i arbeid og aktivitet, med mer.",
 
-  "skrivtiloss.familieogbarn.lenke.tittel": "Familie og barn",
+  "skrivtiloss.familieogbarn.lenke.tittel": "Familie, barn og familie",
   "skrivtiloss.familieogbarn.lenke.ingress":
     "Foreldrepenger, overgangsstønad, kontantstøtte, barnetrygd, barnebidrag og pleiepenger med mer.",
 
   "skrivtiloss.pensjonist.lenke.tittel": "Pensjonist",
   "skrivtiloss.pensjonist.lenke.ingress":
-    "Alderspensjon, barnepensjon, etterlattepensjon og krigspensjon med mer.",
+    "Alderspensjon, barnepensjon, etterlattepensjon og krigspensjon, med mer.",
 
   "skrivtiloss.ufor.lenke.tittel": "Ufør",
   "skrivtiloss.ufor.lenke.ingress":
-    "Uføretrygd, barnetillegg og gjenlevendetillegg med mer.",
+    "Uføretrygd, barnetillegg og gjenlevendetillegg, med mer.",
 
   "skrivtiloss.hjelpemidler.lenke.tittel": "Hjelpemidler",
   "skrivtiloss.hjelpemidler.lenke.ingress":
@@ -68,8 +69,8 @@ export default {
 
   "skrivtiloss.hjelpemidler.tittel": "Skriv til oss om hjelpemidler",
   "skrivtiloss.hjelpemidler.description":
-    "Send spørsmål eller opplysninger om tekniske hjelpemidler, ortopediske hjelpemidler, bilstønad, tolk, tilrettelegging og tilpasning.",
-  "skrivtiloss.hjelpemidler.infotekst": " ",
+    "Send spørsmål eller opplysninger om tekniske hjelpemidler, ortopediske hjelpemidler, bilstønad, tolk," +
+    " tilrettelegging og tilpasning.",
   "skrivtiloss.hjelpemidler.generelt.tittel": "Generelle spørsmål",
   "skrivtiloss.hjelpemidler.generelt.ingress":
     "Gjelder spørsmål om hjelpemidler eller tiltak som bidrar til å reduserer praktiske problemer for personer" +
@@ -121,7 +122,8 @@ export default {
     "Arbeidssøker, dagpenger, arbeidsavklaringspenger eller sykemeldt",
   "kontaktoss.ringoss.tastevalg.tast3.tittel": "Tast 3",
   "kontaktoss.ringoss.tastevalg.tast3.beskrivelse":
-    "Foreldrepenger, engangsstønad, svangerskapspenger, barnebidrag, barnetrygd, kontantstøtte, enslig forsørger, pleiepenger og andre familieytelser",
+    "Foreldrepenger, engangsstønad, svangerskapspenger, barnebidrag, barnetrygd, kontantstøtte, enslig forsørger," +
+    " pleiepenger og andre familieytelser",
   "kontaktoss.ringoss.tastevalg.tast4.tittel": "Tast 4",
   "kontaktoss.ringoss.tastevalg.tast4.beskrivelse": "Uføretrygd og honnørkort",
   "kontaktoss.ringoss.tastevalg.tast5.tittel": "Tast 5",
@@ -163,7 +165,8 @@ export default {
 
   "kontaktoss.klage.tittel": "Klage og tilbakemeldinger",
   "kontaktoss.klage.beskrivelse":
-    "Du kan klage hvis du har fått helt eller delvis avslag på søknaden din, og hvis du har hatt en negativ opplevelse i møte med NAV",
+    "Du kan klage hvis du har fått helt eller delvis avslag på søknaden din, og hvis du har hatt en negativ" +
+    " opplevelse i møte med NAV",
   "kontaktoss.klage.knapp": "Slik går du frem for å klage",
 
   "kontaktoss.tolketjenesten.tittel": "Kontakt tolketjenesten",
@@ -184,7 +187,8 @@ export default {
 
   "tilbakemeldinger.klageanke.tittel": "Klage eller anke på vedtak",
   "tilbakemeldinger.klageanke.beskrivelse":
-    "Har du fått helt eller delvis avslag på søknaden din, kan du klage på avslaget. Får du ikke medhold i klagen din i NAV, kan du med noen unntak anke til Trygderetten.",
+    "Har du fått helt eller delvis avslag på søknaden din, kan du klage på avslaget. Får du ikke medhold i klagen din" +
+    " i NAV, kan du med noen unntak anke til Trygderetten.",
   "tilbakemeldinger.klageanke.link": "Send klage eller anke på vedtak",
 
   "tilbakemeldinger.serviceklage.tittel": "Klage på service",
@@ -206,7 +210,8 @@ export default {
 
   "tilbakemeldinger.feilogmangler.tittel": "Feil og mangler på nav.no",
   "tilbakemeldinger.feilogmangler.beskrivelse":
-    "Har du oppdaget en feil på nav.no? Kanskje en teknisk feil, feil informasjon eller for lav grad av universell utforming? Da vil vi gjerne høre fra deg.",
+    "Har du oppdaget en feil på nav.no? Kanskje en teknisk feil, feil informasjon eller for lav grad av universell" +
+    " utforming? Da vil vi gjerne høre fra deg.",
   "tilbakemeldinger.feilogmangler.link": "Meld fra om feil og mangler",
   "tilbakemeldinger.feilogmangler.form.tittel": "Feil og mangler",
   "tilbakemeldinger.feilogmangler.form.veileder":
@@ -312,10 +317,12 @@ export default {
 
   "seo.kontaktoss.tittel": "Kontakt oss - www.nav.no",
   "seo.kontaktoss.description":
-    "Kontakt NAV via chat, skriv til oss, telefon eller sosiale medier. Finn ditt NAV-kontor, kontakt tolketjenesten, send klage eller tilbakemeldinger.",
+    "Kontakt NAV via chat, skriv til oss, telefon eller sosiale medier. Finn ditt NAV-kontor, kontakt tolketjenesten," +
+    " send klage eller tilbakemeldinger.",
   "seo.tilbakemeldinger.tittel": "Tilbakemeldinger til NAV - www.nav.no",
   "seo.tilbakemeldinger.description":
-    "Her kan du gi tilbakemeldinger til NAV, klage eller anke på vedtak, service og melde fra om feil og mangler på nav.no. Du kan også gi oss ros om våre tjenester og medarbeidere.",
+    "Her kan du gi tilbakemeldinger til NAV, klage eller anke på vedtak, service og melde fra om feil og mangler på" +
+    " nav.no. Du kan også gi oss ros om våre tjenester og medarbeidere.",
   "seo.rostilnav.tittel": "Ros til NAV - www.nav.no",
   "seo.rostilnav.description":
     "Har du en god opplevelse? Vil du rose en medarbeider, eller er det noe annet positivt du vil dele med NAV?",
@@ -331,21 +338,6 @@ export default {
     "Vi hjelper deg med å finne frem på nav.no og kan svare deg på spørsmål om hva du kan ha rett på. Vi kan ikke se" +
     " saken din eller utgi personlige opplysninger.",
 
-  "chat.familie.lenke.tittel": "Foreldre, barn og familie",
-  "chat.familie.lenke.ingress":
-    "Foreldrepenger, overgangsstønad, kontantstøtte, barnetrygd, barnebidrag, pleiepenger med mer.",
-
-  "chat.aap.lenke.tittel": "Jobbsøker",
-  "chat.aap.lenke.ingress":
-    "Dagpenger, arbeidsavklaringspenger, tiltakspenger, tiltak eller veiledning for å komme i jobb eller aktivitet.",
-
-  "chat.sosialhjelp.lenke.tittel": "Sosiale tjenester og økonomisk sosialhjelp",
-  "chat.sosialhjelp.lenke.ingress":
-    "Økonomisk sosialhjelp, midlertidig botilbud eller andre sosiale tjenester.",
-
-  "chat.okonomi.lenke.tittel": "Økonomi og gjeldsrådgivning",
-  "chat.okonomi.lenke.ingress": "Budsjett, økonomi og gjeldsrådgivning.",
-
   "chat.startknapp": "Start chat",
   "chat.advarsel.personvern":
     "<p>Alle samtaler blir lagret anonymt i 6 måneder.</p><p>Ikke oppgi navn, fødselsnummer eller" +
@@ -356,13 +348,16 @@ export default {
     "Foreldrepenger, overgangsstønad, kontantstøtte, barnetrygd, barnebidrag, pleiepenger med mer.",
   "chat.familie.ingress":
     "<p>Foreldrepenger, overgangsstønad, kontantstøtte, barnetrygd, barnebidrag, pleiepenger med mer.</p>" +
+    "<p>Vår chatrobot Frida sitter klar til å svare deg på generelle spørsmål.</p>" +
     "<p>Veileder er tilgjengelig hverdager fra kl. 09.00 til 14.30.</p>",
 
   "chat.aap.tittel": "Jobbsøker",
   "chat.aap.ingressKort":
     "Dagpenger, arbeidsavklaringspenger, tiltakspenger, tiltak eller veiledning for å komme i jobb eller aktivitet.",
   "chat.aap.ingress":
-    "<p>Dagpenger, arbeidsavklaringspenger, tiltakspenger, tiltak eller veiledning for å komme i jobb eller aktivitet.</p>",
+    "<p>Dagpenger, arbeidsavklaringspenger, tiltakspenger, tiltak eller veiledning for å komme i jobb eller" +
+    " aktivitet.</p><p>Vår chatrobot Frida sitter klar til å svare deg på generelle spørsmål.</p>" +
+    "<p>Veileder er tilgjengelig hverdager fra kl. 09.00 til 14.30.</p>",
 
   "chat.sosialhjelp.tittel": "Sosiale tjenester og økonomisk sosialhjelp",
   "chat.sosialhjelp.ingressKort":
