+++ src/pages/skriv-til-oss/SkrivTilOssBase.tsx
@@ -1,18 +1,9 @@
-import React, { useEffect, useState } from "react";
-import { Normaltekst, Sidetittel } from "nav-frontend-typografi";
+import React, { useEffect } from "react";
+import { Sidetittel } from "nav-frontend-typografi";
 import { FormattedMessage, useIntl } from "react-intl";
-import { vars } from "../../Config";
-import { Features, getFeatureToggleStatusMultiple } from "../../utils/unleash";
-import NavFrontendSpinner from "nav-frontend-spinner";
-import AlertStripe from "nav-frontend-alertstriper";
 import { LenkepanelData } from "../../types/lenker";
 import SkrivTilOssLenkepanel from "./SkrivTilOssLenkepanel";
 import BreadcrumbsWrapper from "../../components/breadcrumbs/BreadcrumbsWrapper";
-
-const enabledName = vars.unleash.skrivTilOssEnabledName;
-const svartidName = vars.unleash.langSvartidName;
-const enabledDefault = vars.unleash.skrivTilOssEnabledDefault;
-const svartidDefault = vars.unleash.langSvartidDefault;
 
 const cssPrefix = "skriv-til-oss";
 
@@ -30,40 +21,6 @@ const SkrivTilOssBase = ({ tittel, children, lenker }: Props) => {
     document.title = documentTitle;
   }, [documentTitle]);
 
-  const [unleashResponded, setUnleashResponded] = useState(false);
-  const [langSvartid, setLangSvartid] = useState(svartidDefault);
-  const [skrivTilOssEnabled, setSkrivTilOssEnabled] = useState(enabledDefault);
-
-  const unleashTogglesResponse = (unleashToggles: Features, error: any) => {
-    setUnleashResponded(true);
-    if (error) {
-      console.log(`Unleash error: ${error}`);
-      return;
-    }
-
-    setLangSvartid(unleashToggles[svartidName]);
-    setSkrivTilOssEnabled(unleashToggles[enabledName]);
-  };
-
-  useEffect(() => {
-    getFeatureToggleStatusMultiple(
-      [svartidName, enabledName],
-      unleashTogglesResponse
-    );
-  }, []);
-
-  if (!unleashResponded) {
-    return <NavFrontendSpinner negativ={true} />;
-  }
-
-  if (!skrivTilOssEnabled) {
-    return (
-      <AlertStripe type="advarsel">
-        <FormattedMessage id={"skrivtiloss.disabled"} />
-      </AlertStripe>
-    );
-  }
-
   return (
     <div className={`${cssPrefix} pagecontent`}>
       <BreadcrumbsWrapper />
@@ -73,13 +30,6 @@ const SkrivTilOssBase = ({ tittel, children, lenker }: Props) => {
         </Sidetittel>
       </div>
       <div className={`${cssPrefix}__ingress`}>
-        <Normaltekst className={`${cssPrefix}__svartid`}>
-          <FormattedMessage
-            id={"skrivtiloss.svartid"}
-            values={{ numDager: vars.svartidDager }}
-          />
-          {langSvartid && <FormattedMessage id={"skrivtiloss.svartid.lang"} />}
-        </Normaltekst>
         {children}
       </div>
       {lenker && (
