+++ src/pages/skriv-til-oss/sider/SkrivTilOssForside.tsx
@@ -1,14 +1,57 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 import MetaTags from "react-meta-tags";
 
 import SkrivTilOssBase from "../SkrivTilOssBase";
 import { LenkepanelData } from "types/lenker";
 import { Normaltekst } from "nav-frontend-typografi";
-import { urls } from "../../../Config";
+import { urls, vars } from "../../../Config";
+import { Features, getFeatureToggleStatusMultiple } from "../../../utils/unleash";
+import NavFrontendSpinner from "nav-frontend-spinner";
+import AlertStripe from "nav-frontend-alertstriper";
+
+const svartidName = vars.unleash.langSvartidName;
+const svartidDefault = vars.unleash.langSvartidDefault;
+const enabledName = vars.unleash.skrivTilOssEnabledName;
+const enabledDefault = vars.unleash.skrivTilOssEnabledDefault;
 
 const Ingress = () => {
   const intl = useIntl();
+
+  const [unleashResponded, setUnleashResponded] = useState(false);
+  const [langSvartid, setLangSvartid] = useState(svartidDefault);
+  const [skrivTilOssEnabled, setSkrivTilOssEnabled] = useState(enabledDefault);
+
+  const unleashTogglesResponse = (unleashToggles: Features, error: any) => {
+    setUnleashResponded(true);
+    if (error) {
+      console.log(`Unleash error: ${error}`);
+      return;
+    }
+
+    setLangSvartid(unleashToggles[svartidName]);
+    setSkrivTilOssEnabled(unleashToggles[enabledName]);
+  };
+
+  useEffect(() => {
+    getFeatureToggleStatusMultiple(
+      [svartidName, enabledName],
+      unleashTogglesResponse
+    );
+  }, []);
+
+  if (!unleashResponded) {
+    return <NavFrontendSpinner negativ={true} />;
+  }
+
+  if (!skrivTilOssEnabled) {
+    return (
+      <AlertStripe type="advarsel">
+        <FormattedMessage id={"skrivtiloss.disabled"} />
+      </AlertStripe>
+    );
+  }
+
   return (
     <>
       <MetaTags>
@@ -18,7 +61,14 @@ const Ingress = () => {
           content={intl.messages["skrivtiloss.description"] as string}
         />
       </MetaTags>
-      <Normaltekst className="skriv-til-oss__infotekst">
+      <Normaltekst>
+        <FormattedMessage
+          id={"skrivtiloss.svartid"}
+          values={{ numDager: vars.svartidDager }}
+        />
+        {langSvartid && <FormattedMessage id={"skrivtiloss.svartid.lang"} />}
+      </Normaltekst>
+      <Normaltekst>
         <FormattedMessage id="skrivtiloss.ingress" />
       </Normaltekst>
     </>
@@ -61,12 +111,12 @@ const lenker: LenkepanelData[] = [
     ingress: <FormattedMessage id={"skrivtiloss.hjelpemidler.lenke.ingress"} />,
     url: urls.skrivTilOss.hjelpemidler,
   },
-  {
-    tittel: "skrivtiloss.sosial.lenke.tittel",
-    ingress: <FormattedMessage id={"skrivtiloss.sosial.lenke.ingress"} />,
-    url: urls.skrivTilOss.sosial,
-    external: true,
-  }
+  // {
+  //   tittel: "skrivtiloss.sosial.lenke.tittel",
+  //   ingress: <FormattedMessage id={"skrivtiloss.sosial.lenke.ingress"} />,
+  //   url: urls.skrivTilOss.sosial,
+  //   external: true,
+  // }
 ];
 
 const SkrivTilOssForside = () => (
